Add files via upload
This is a macro/python script. You must install Visual Studio (2019), Anaconda, and Excel.

XLwings comes with Anaconda, to start it open the Anaconda terminal and type 'xlwings addin install' to activate it in Excel. Once this is set up, make sure that the excel file and python script are in the same folder with the same name (i.e. 'contact_finder' (folder) 'contact_finder.xlsm' / 'contact_finder.py'). Reach out with questions!